Final Project
For my final project, my goal was to create a tool to compare dataset statistics. My tool takes two feature classes as input, along with a field name. It checks whether the field is numeric and exists in both feature classes. If both these conditions are met, the tool calculates statistics for that field for both feature classes. Statistics include sum, average, standard deviation, minimum, and maximum. Finally, the tool reports a comparison between the datasets.